rabbitmq-log-message-queue/requirements.txt
pika==1.3.2
click==8.1.7
rich==13.9.4
python-dotenv==1.0.1
pyyaml==6.0.2
pytest==8.3.4
requests==2.31.0
